djangoapp/requirements.txt
# Core
Django>=4.2.1,<4.3
psycopg2-binary>=2.9.6,<2.10
python-dotenv>=1.0.0,<1.1
requests==2.31.0

# Imagens (Essencial para as fotos das peças)
Pillow>=10.0.0

# UI e Formulários Bonitos
django-crispy-forms==2.5
crispy-bootstrap5>=2024.10

# Editor de Texto e Segurança
django-summernote>=0.8.20.0,<0.8.21
django-axes>=6.0.1,<6.1
